rename scope function and add some comments
