+++ kotest-framework/kotest-framework-datatest/src/jvmTest/kotlin/io/kotest/datatest/UnstableTestNameWithLeafIsolationTest.kt
@@ -1,6 +1,6 @@
 package io.kotest.datatest
 
-import io.kotest.core.config.ExperimentalKotest
+import io.kotest.common.ExperimentalKotest
 import io.kotest.core.spec.Isolate
 import io.kotest.core.spec.IsolationMode
 import io.kotest.core.spec.style.DescribeSpec
